\name{move.HMM.dwell.plot}
\alias{move.HMM.dwell.plot}
\title{Dwell Plot}
\usage{
  move.HMM.dwell.plot(move.HMM)
}
\arguments{
  \item{move.HMM}{A move.HMM object containing a fitted HMM
  model.}
}
\value{
  A plot of dwell times
}
\description{
  This function plots the (geometric) pdf of the fitted
  model against the dwell times from the state sequence
  predicted by the Viterbi algorithm.
}

