\name{move.HSMM.dwell.plot}
\alias{move.HSMM.dwell.plot}
\title{Dwell Plot}
\usage{
  move.HSMM.dwell.plot(move.HSMM)
}
\arguments{
  \item{move.HSMM}{A move.HSMM object containing a fitted
  HSMM model.}
}
\value{
  A plot of dwell times
}
\description{
  This function plots the pdf of the fitted model against
  the dwell times from the state sequence predicted by the
  Viterbi algorithm.
}

