\name{stationary}
\alias{stationary}
\title{Calculate the stationary distribution from the t.p.m.}
\usage{
  stationary(gamvect)
}
\arguments{
  \item{gamvect}{A vector containing the MLEs of the
  t.p.m.}
}
\value{
  a vector containing the stationary distribution
}
\description{
  This function, is used to get the numerical derivative
  required to get the covariance matrix of the stationary
  distribution.
}

